Refactoring to reduce duplicated code across tests. Removing timeouts.
