+++ test/functional/example.test.ts
@@ -1,41 +1,15 @@
-import { Browser, Page } from 'puppeteer';
+import { Page } from 'puppeteer';
 import { IndexPage } from '../page-objects/IndexPage';
-import { startAppServer, tearDown } from '../accessibility/config/common';
-
-const puppeteer = require('puppeteer');
-const config = require('config');
-const httpProxy = config.get('httpProxy');
+import { getNewPage, tearDown } from '../utils/common';
 
 describe('Load index page', () => {
   let indexPage: IndexPage;
 
-  async function startBrowser() {
-    const args = ['--no-sandbox', '--start-maximized'];
-    if (httpProxy) {
-      args.push(`-proxy-server=${httpProxy}`);
-    }
-
-    const opts = {
-      args,
-      headless: true,
-      timeout: 10000,
-      ignoreHTTPSErrors: true
-    };
-    return puppeteer.launch(opts);
-  }
-
   before('setup browser', async () => {
-    let page: Page;
-    if (process.env.NODE_ENV === 'development') {
-      const server = await startAppServer();
-      page = server.page;
-    } else {
-      const browser: Browser = await startBrowser();
-      page = await browser.newPage();
-    }
+    const page: Page = await getNewPage();
     indexPage = new IndexPage(page);
     await indexPage.gotoPage();
-    await indexPage.screenshot('start_service');
+    await indexPage.screenshot('functional-screenshots', 'start_service');
   });
 
   after(async () => {
